@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
     defaultConfig {
         applicationId "life.knowledge4.videocroppersample"
         minSdkVersion 16
-        targetSdkVersion 24
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -32,9 +32,9 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     //compile 'life.knowledge4:k4l-video-trimmer:0.1-SNAPSHOT'
     compile project(':k4l-video-trimmer')
-    compile 'com.android.support:appcompat-v7:24.0.0'
+    compile 'androidx.appcompat:appcompat:1.0.2'
     testCompile 'junit:junit:4.12'
-    androidTestCompile 'com.android.support.test.espresso:espresso-core:2.2.2'
-    androidTestCompile 'com.android.support.test:runner:0.5'
-    androidTestCompile 'com.android.support:support-annotations:24.0.0'
+    androidTestCompile 'androidx.test.espresso:espresso-core:3.2.0'
+    androidTestCompile 'androidx.test:runner:1.2.0'
+    androidTestCompile 'androidx.annotation:annotation:1.1.0'
 }
