@@ -26,7 +26,8 @@ package life.knowledge4.videotrimmer.view;
 import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import java.util.List;
 import java.util.Vector;
